modified GoldScriptFiles
골드 관련 스크립트 수정
