+++ Assets/EnemyMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28c5e6541c2e89e49a612d4ef78157c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
