+++ Assets/PreFabs/Enemies/Enemy.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 5277318453832661174}
   - component: {fileID: 5277318453832661175}
   - component: {fileID: 114731223250471600}
+  - component: {fileID: 6330629966176447593}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Enemy
@@ -106,3 +107,15 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   speed: 10
   coin: 100
+--- !u!114 &6330629966176447593
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5277318453832661172}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 28c5e6541c2e89e49a612d4ef78157c6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
